Changes include:
--- 
 r2037 | crath | 2009-03-30 00:45:27 +0200 (Mo, 30. Mär 2009) | 3 lines
   M /branches/nao/src/firevision/tools/firestation/firestation.cpp
   M /branches/nao/src/firevision/tools/yuv_viewer/Makefile
   M /branches/nao/src/firevision/tools/yuv_viewer/main.cpp
   M /branches/nao/src/firevision/tools/yuv_viewer/yuv_viewer_gui.cpp
   M /branches/nao/src/firevision/tools/yuv_viewer/yuv_viewer_gui.h

 Changed documentation and fv prefix
--- 
 r2038 | crath | 2009-03-30 13:31:55 +0200 (Mo, 30. Mär 2009) | 2 lines
   M /branches/nao/src/firevision/fvutils/draw/field.h

 Removing virtual keyword
--- 
 r2043 | tim | 2009-03-31 11:41:36 +0200 (Di, 31. Mär 2009) | 7 lines
   M /branches/nao/src/libs/netcomm/utils/ntoh64.h

 Compile fix for optimized compilation



git-svn-id: https://lechuck.informatik.rwth-aachen.de:2342/svn/fawkes/trunk@2071 fc82b16a-671c-0410-8732-9903a8d2dd0d

+++ src/libs/netcomm/utils/ntoh64.h
@@ -29,8 +29,8 @@
 #include <stdint.h>
 #include <arpa/inet.h>
 
-#define __ntoh64(x) (((uint64_t)(ntohl((uint32_t)((x << 32) >> 32))) << 32) | \
-		       (uint32_t)ntohl(((uint32_t)(x >> 32))))
+#define __ntoh64(x) (((uint64_t)(ntohl((uint32_t)(((uint64_t)x << 32) >> 32))) << 32) | \
+		       (uint32_t)ntohl(((uint32_t)((uint64_t)x >> 32))))
 
 #define __hton64(x) ntohll(x)
